feat(components): move DataAnalytics to components
